+++ sprite.hpp
@@ -18,7 +18,7 @@ public:
 	 *
 	 * returns: sprite handle.
 	 */
-	Sprite(Texture texture, Vector position, int z, Vector scale, float rotation, Color color);
+	Sprite(Texture texture, Vector position, int z, Vector scale, double rotation, Color color);
 
 	class Loader {
 	private:
@@ -26,7 +26,7 @@ public:
 		Vector pposition;
 		int pz{0};
 		Vector pscale{1};
-		float protation{0};
+		double protation{0};
 		Color pcolor{1};
 
 	public:
@@ -35,7 +35,7 @@ public:
 		Loader& position(Vector sv) { pposition = sv; return *this; }
 		Loader& z(int i) { pz = i; return *this; }
 		Loader& scale(Vector sv) { pscale = sv; return *this; }
-		Loader& rotation(float f) { protation = f; return *this; }
+		Loader& rotation(double f) { protation = f; return *this; }
 		Loader& color(Color sc) { pcolor = sc; return *this; }
 
 		Sprite* load() {
